pyproject.toml
[tool.poetry]
name = "reagency"
version = "0.1.0"
description = "lightweight abstractions for building agent scaffolds"
authors = ["danielreuter <dreuter14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langfuse = "^2.52.1"
mypy = "^1.12.1"
openai = "^1.52.1"
anthropic = "^0.37.1"
petname = "^2.6"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
pillow = "^11.0.0"
joblib = "^1.4.2"
pytest-asyncio = "^0.24.0"
nest-asyncio = "^1.6.0"
numpy = "^2.1.2"
torch = "^2.5.0"
einops = "^0.8.0"
tqdm = "^4.66.6"
agentlens = "0.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
